Multiplayer fixes for spells, status text, progression abilities
